+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 Roadmap
 =======
 
+- TODO subtype combinator: add a reference to the super type prototype, fix #22
+- TODO add `Dict(A, B)` combinator, for dictionaries, fix #21
+
 # v.0.0.12
 
+- more verbose error messages, fix #25
+- list#is() depends on `this`, fix #27
+- tuple#is() depends on `this`, fix #28
 - `enums.of()` should generate values equals to keys, fix #23
 
 # v.0.0.11